+++ examples/mmada/training/train_mmada.py
@@ -23,7 +23,9 @@ from PIL import Image
 from training.data import Text2ImageDataset
 from training.imagenet_dataset import ImageNetDataset
 from training.prompting_utils import UniversalPrompting
+from training.utils import AverageMeter, get_config, image_transform, mask_or_random_replace_tokens
 from transformers import AutoConfig, AutoTokenizer
+from utils.ckpt import init_from_ckpt
 from utils.train_step import TrainStepMmaDA, do_ckpt_combine_online, prepare_train_network
 
 import mindspore as ms
@@ -34,11 +36,9 @@ from mindspore.mint.distributed import get_rank, get_world_size, init_process_gr
 
 from mindone.diffusers.models.model_loading_utils import load_checkpoint_and_dispatch
 from mindone.diffusers.training_utils import pynative_no_grad, set_seed
-from utils import init_from_ckpt
 
 SYSTEM_PROMPT_LEN = 28
 
-from training.utils import AverageMeter, get_config, image_transform, mask_or_random_replace_tokens
 
 logger = logging.getLogger(__name__)
